+++ changes.md
@@ -3,6 +3,7 @@
 ## Changes in 4.5.x
 1. Fix - position of TFrame in canvas/pad
 2. Fix - use histogram fMinimum/fMaximum when creating color palette
+3. Fix - correctly draw empty th2 bins when zmin<0 is specified  
 
 
 ## Changes in 4.5.1